feat(core): move ECS accessors to accessors.hpp
