+++ core/CMakeLists.txt
@@ -166,6 +166,7 @@ set(CUBOS_CORE_INCLUDE
     "include/cubos/core/ecs/event_pipe.hpp"
     "include/cubos/core/ecs/event_reader.hpp"
     "include/cubos/core/ecs/event_writer.hpp"
+    "include/cubos/core/ecs/accessors.hpp"
 )
 
 # Create core library
